requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
python-docx
openai
sentence-transformers
faiss-cpu
chromadb
langchain
faiss-cpu
chromadb
numpy
google-generativeai
python-dotenv

